# src/http/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TMDB M-Team Hub Management API",
    "version": "1.0.0"
  },
  "security": [
    {
      "ManagementSession": []
    }
  ],
  "paths": {
...
    "/api/douban/search": {
      "get": {
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 20,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Douban search results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoubanSearchResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid Douban search request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/ManagementUnauthorized"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "502": {
            "description": "Douban request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/api/douban/library": {
      "get": {
        "parameters": [
          {
            "name": "force_refresh",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1200,
              "default": 200
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Current Douban wish and collect library",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoubanLibraryResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Douban authentication is not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/ManagementUnauthorized"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "502": {
            "description": "Douban request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/api/douban/tags": {
      "get": {
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1200,
              "default": 80
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Configured Douban wanted-tag history",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoubanTagHistoryResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Douban authentication is not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/ManagementUnauthorized"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/api/douban/subject/{id}": {
      "get": {
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Normalized Douban subject detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoubanSubjectDetailDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid Douban subject",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/ManagementUnauthorized"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "502": {
            "description": "Douban request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/api/douban/subject/{id}/interest": {
      "post": {
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DoubanInterestRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated Douban interest",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoubanInterestResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid interest mutation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/ManagementUnauthorized"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "502": {
            "description": "Douban request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/api/douban/image": {
      "get": {
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Proxied Douban image bytes",
            "content": {
              "image/jpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "Invalid image URL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/ManagementUnauthorized"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "502": {
            "description": "Douban image request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/api/douban/qr/start": {
      "post": {
        "responses": {
          "200": {
            "description": "Started an isolated Douban QR login session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoubanQrStartResponseDto"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/ManagementUnauthorized"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "502": {
            "description": "Douban QR request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/api/douban/qr/poll": {
      "get": {
        "parameters": [
          {
            "name": "session_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Current QR login status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoubanQrPollResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "QR session is missing or expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/ManagementUnauthorized"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "502": {
            "description": "Douban QR request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/api/douban/qr/image": {
      "get": {
        "parameters": [
          {
            "name": "session_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "QR session image",
            "content": {
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "QR session is missing or expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/ManagementUnauthorized"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ManagementSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "tmdb_mteam_admin_session"
      }
    },
    "responses": {
      "ManagementUnauthorized": {
        "description": "Management authentication is required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorDto"
            }
          }
        }
      },
      "MethodNotAllowed": {
        "description": "HTTP method is not allowed for this endpoint",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorDto"
            }
          }
        }
      },
      "InternalError": {
        "description": "Internal server error with a redacted public message",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorDto"
            }
          }
        }
      },
      "ServiceUnavailable": {
        "description": "A required local service or store is temporarily unavailable",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorDto"
            }
          }
        }
      }
    },
    "schemas": {
...
      "DoubanRatingDto": {
        "type": "object",
        "additionalProperties": false,
        "required": ["value", "count", "info", "star_count"],
        "properties": {
          "value": {
            "type": ["number", "null"]
          },
          "count": {
            "type": ["integer", "null"],
            "minimum": 0
          },
          "info": {
            "type": "string"
          },
          "star_count": {
            "type": ["number", "null"]
          }
        }
      },
      "DoubanSearchItemDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "source",
          "media_type",
          "id",
          "subject_id",
          "title",
          "url",
          "abstract_text",
          "abstract_2",
          "cover_url",
          "poster_url",
          "rating",
          "vote_average"
        ],
        "properties": {
          "source": {
            "type": "string",
            "const": "douban"
          },
          "media_type": {
            "type": "string",
            "const": "douban"
          },
          "id": {
            "type": "string"
          },
          "subject_id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "abstract_text": {
            "type": "string"
          },
          "abstract_2": {
            "type": "string"
          },
          "cover_url": {
            "type": "string"
          },
          "poster_url": {
            "type": "string"
          },
          "rating": {
            "$ref": "#/components/schemas/DoubanRatingDto"
          },
          "vote_average": {
            "type": ["number", "null"]
          }
        }
      },
      "DoubanSearchResponseDto": {
        "type": "object",
        "additionalProperties": false,
        "required": ["items", "page", "page_size", "has_more"],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DoubanSearchItemDto"
            }
          },
          "page": {
            "type": "integer",
            "minimum": 1
          },
          "page_size": {
            "type": "integer",
            "minimum": 1,
            "maximum": 20
          },
          "has_more": {
            "type": "boolean"
          }
        }
      },
      "DoubanSubjectDetailDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "source",
          "media_type",
          "id",
          "subject_id",
          "url",
          "title",
          "image",
          "poster_url",
          "directors",
          "writers",
          "actors",
          "genres",
          "date_published",
          "duration",
          "summary",
          "rating"
        ],
        "properties": {
          "source": {
            "type": "string",
            "const": "douban"
          },
          "media_type": {
            "type": "string",
            "const": "douban"
          },
          "id": {
            "type": "string"
          },
          "subject_id": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "original_title": {
            "type": "string"
          },
          "aka": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "languages": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "countries": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "image": {
            "type": "string"
          },
          "poster_url": {
            "type": "string"
          },
          "directors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "writers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "actors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "genres": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "date_published": {
            "type": "string"
          },
          "duration": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "rating": {
            "$ref": "#/components/schemas/DoubanRatingDto"
          },
          "user_interest": {
            "type": "string",
            "enum": ["wish", "collect"]
          },
          "user_rating": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5
          }
        }
      },
      "DoubanInterestRequestDto": {
        "type": "object",
        "additionalProperties": false,
        "required": ["interest"],
        "properties": {
          "interest": {
            "type": "string",
            "enum": ["wish", "collect"]
          },
          "rating": {
            "type": ["integer", "null"],
            "minimum": 1,
            "maximum": 5
          },
          "tags": {
            "type": "string"
          }
        }
      },
      "DoubanInterestResponseDto": {
        "type": "object",
        "additionalProperties": false,
        "required": ["ok", "interest", "rating", "tags"],
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "interest": {
            "type": "string",
            "enum": ["wish", "collect"]
          },
          "rating": {
            "type": ["integer", "null"],
            "minimum": 1,
            "maximum": 5
          },
          "tags": {
            "type": "string"
          }
        }
      },
      "DoubanSnapshotCompleteness": {
        "type": "string",
        "enum": ["complete", "partial"]
      },
      "DoubanLibraryItemDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "source",
          "media_type",
          "id",
          "subject_id",
          "title",
          "url",
          "abstract_text",
          "abstract_2",
          "cover_url",
          "poster_url",
          "status",
          "status_label",
          "date",
          "comment",
          "tags",
          "user_rating"
        ],
        "properties": {
          "source": {
            "type": "string",
            "const": "douban"
          },
          "media_type": {
            "type": "string",
            "const": "douban"
          },
          "id": {
            "type": "string"
          },
          "subject_id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "abstract_text": {
            "type": "string"
          },
          "abstract_2": {
            "type": "string"
          },
          "cover_url": {
            "type": "string"
          },
          "poster_url": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["wish", "collect"]
          },
          "status_label": {
            "type": "string"
          },
          "date": {
            "type": "string"
          },
          "comment": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "user_rating": {
            "type": ["integer", "null"],
            "minimum": 1,
            "maximum": 5
          }
        }
      },
      "DoubanLibraryListDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "status",
          "label",
          "items",
          "completeness",
          "fetched_pages",
          "truncated_by_limit",
          "end_observed"
        ],
        "properties": {
          "status": {
            "type": "string",
            "enum": ["wish", "collect"]
          },
          "label": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DoubanLibraryItemDto"
            }
          },
          "completeness": {
            "$ref": "#/components/schemas/DoubanSnapshotCompleteness"
          },
          "fetched_pages": {
            "type": "integer",
            "minimum": 0
          },
          "truncated_by_limit": {
            "type": "boolean"
          },
          "end_observed": {
            "type": "boolean"
          }
        }
      },
      "DoubanLibraryResponseDto": {
        "type": "object",
        "additionalProperties": false,
        "required": ["source", "cached", "fetched_at", "ttl_seconds", "limit", "wish", "collect"],
        "properties": {
          "source": {
            "type": "string",
            "const": "douban"
          },
          "cached": {
            "type": "boolean"
          },
          "fetched_at": {
            "type": "integer",
            "minimum": 0
          },
          "ttl_seconds": {
            "type": "integer",
            "minimum": 0
          },
          "limit": {
            "type": "integer",
            "minimum": 1,
            "maximum": 1200
          },
          "wish": {
            "$ref": "#/components/schemas/DoubanLibraryListDto"
          },
          "collect": {
            "$ref": "#/components/schemas/DoubanLibraryListDto"
          }
        }
      },
      "DoubanTagCountDto": {
        "type": "object",
        "additionalProperties": false,
        "required": ["tag", "count", "category"],
        "properties": {
          "tag": {
            "type": "string"
          },
          "count": {
            "type": "integer",
            "minimum": 0
          },
          "category": {
            "type": "string"
          }
        }
      },
      "DoubanTagCategoryDto": {
        "type": "object",
        "additionalProperties": false,
        "required": ["name", "wanted_tag"],
        "properties": {
          "name": {
            "type": "string"
          },
          "wanted_tag": {
            "type": "string"
          }
        }
      },
      "DoubanTagHistoryResponseDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "source",
          "cached",
          "updated_at",
          "tags",
          "tag_counts",
          "subscription_categories"
        ],
        "properties": {
          "source": {
            "type": "string"
          },
          "cached": {
            "type": "boolean"
          },
          "updated_at": {
            "type": ["integer", "null"],
            "minimum": 0
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "tag_counts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DoubanTagCountDto"
            }
          },
          "subscription_categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DoubanTagCategoryDto"
            }
          }
        }
      },
      "DoubanQrStartResponseDto": {
        "type": "object",
        "additionalProperties": false,
        "required": ["session_id", "image_url"],
        "properties": {
          "session_id": {
            "type": "string"
          },
          "image_url": {
            "type": "string"
          }
        }
      },
      "DoubanQrPollResponseDto": {
        "type": "object",
        "additionalProperties": false,
        "required": ["done", "login_status", "message", "description", "cookie_saved"],
        "properties": {
          "done": {
            "type": "boolean"
          },
          "login_status": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "cookie_saved": {
            "type": "boolean"
          }
        }
      },
      "ApiErrorDto": {
        "type": "object",
        "additionalProperties": false,
        "required": ["code", "message"],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      }
    }
  }
}